+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 3.0.80
 
+* Update fontawesome CDN to version 6.4.2
+* Remove default libraries such as bootstrap, datatable, jquery, jquery UI
 * Add controller for managing custom libraries
 * New model for storing links to libraries like JQuery
 * Fix bug where publishing a nav or side nav publishes all pages (including drafts)